+++ src/euphorie/upgrade/client/v20/v20260630154225_install_the_new_registry_record_to_customize_the_display_of_the_client/upgrade.py
@@ -1,0 +1,8 @@
+from collective.ftw.upgrade import UpgradeStep
+
+
+class InstallTheNewRegistryRecordToCustomizeTheDisplayOfTheClient(UpgradeStep):
+    """Install the new registry record to customize the display of the client."""
+
+    def __call__(self):
+        self.install_upgrade_profile()