+++ src/euphorie/upgrade/deployment/v1/20220720110538_upgrade_to_alembic_head/upgrade.py
@@ -1,9 +1,0 @@
-from euphorie.deployment.upgrade.utils import alembic_upgrade
-from ftw.upgrade import UpgradeStep
-
-
-class UpgradeToAlembicHead(UpgradeStep):
-    """Create the OrganisationMembership table."""
-
-    def __call__(self):
-        alembic_upgrade()