+++ src/euphorie/upgrade/deployment/v1/20230421104953_add_use_consultancy_phase_feature_switch/upgrade.py
@@ -1,8 +1,0 @@
-from ftw.upgrade import UpgradeStep
-
-
-class AddUseConsultancyPhaseFeatureSwitch(UpgradeStep):
-    """Add use_consultancy_phase feature switch."""
-
-    def __call__(self):
-        self.install_upgrade_profile()