+++ src/euphorie/upgrade/deployment/v1/20230421141443_add_consultancy_table/upgrade.py
@@ -1,9 +1,0 @@
-from euphorie.deployment.upgrade.utils import alembic_upgrade_to
-from ftw.upgrade import UpgradeStep
-
-
-class AddConsultancyTable(UpgradeStep):
-    """Add consultancy table."""
-
-    def __call__(self):
-        alembic_upgrade_to(self.target_version)