+++ src/euphorie/upgrade/deployment/v1/20230515143459_allow_consultant_deletion/upgrade.py
@@ -1,9 +1,0 @@
-from euphorie.deployment.upgrade.utils import alembic_upgrade_to
-from ftw.upgrade import UpgradeStep
-
-
-class AllowConsultantDeletion(UpgradeStep):
-    """Allow consultant deletion."""
-
-    def __call__(self):
-        alembic_upgrade_to(self.target_version)