+++ src/euphorie/upgrade/deployment/v1/20230606094424_remove_use_consultancy_phase_feature_switch/upgrade.py
@@ -1,8 +1,0 @@
-from ftw.upgrade import UpgradeStep
-
-
-class RemoveUseConsultancyPhaseFeatureSwitch(UpgradeStep):
-    """Remove use_consultancy_phase feature switch."""
-
-    def __call__(self):
-        self.install_upgrade_profile()