+++ src/euphorie/upgrade/deployment/v1/20231128125729_notifications_registry_entries/upgrade.py
@@ -1,8 +1,0 @@
-from ftw.upgrade import UpgradeStep
-
-
-class NotificationsRegistryEntries(UpgradeStep):
-    """Notifications registry entries."""
-
-    def __call__(self):
-        self.install_upgrade_profile()