+++ src/euphorie/upgrade/deployment/v20/v20260702121717_intall_collective_ftw_upgrade/upgrade.py
@@ -1,0 +1,8 @@
+from collective.ftw.upgrade import UpgradeStep
+
+
+class IntallCollectiveFtwUpgrade(UpgradeStep):
+    """Intall collective.ftw.upgrade."""
+
+    def __call__(self):
+        self.ensure_profile_installed("collective.ftw.upgrade:default")